+++ analytics_toolkit/sql/dml/__init__.py
@@ -3,8 +3,6 @@ from .io.gp_cancel import gp_cancel_all_running_queries
 from .io.read_sql import read_sql
 from .load.load_df import build_load_df_plan, load_df
 from .table import (
-    build_gp_create_many_partitions_sqls,
-    create_table_from_sql,
     drop_many_partitions,
     gp_create_many_partitions,
     gp_vacuum,
@@ -12,10 +10,8 @@ from .table import (
 from .transfer.flow.api import build_transfer_table_plan, transfer_table
 
 __all__ = [
-    "build_gp_create_many_partitions_sqls",
     "build_load_df_plan",
     "build_transfer_table_plan",
-    "create_table_from_sql",
     "drop_many_partitions",
     "execute_sql",
     "gp_create_many_partitions",