+++ analytics_toolkit/sql/dml/table/api.py
@@ -47,7 +47,6 @@ from .partitions import (
     build_drop_many_partitions_plan,
     build_drop_many_partitions_sqls,
     build_gp_create_many_partitions_plan,
-    build_gp_create_many_partitions_sqls,
     drop_many_partitions,
     gp_create_many_partitions,
 )